.requirements.txt
numpy
pandas
jupyterlab
PyYAML
Sphinx
sphinx-rtd-theme
sphinxcontrib-jsmath
sphinxcontrib-napoleon
sphinxcontrib-websupport
sphinx-copybutton
pytest
pytest-cov
pytest-reportlog
pytest-html
pylint
pylint-report
